refactor: extract usePromptSegmentsSync to satisfy 300-line limit
